+++ src/main/java/szewek/mcflux/items/ItemFESniffer.java
@@ -2,7 +2,6 @@ package szewek.mcflux.items;
 
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumActionResult;
@@ -16,7 +15,7 @@ import szewek.mcflux.api.fe.FE;
 import szewek.mcflux.api.fe.IFlavorEnergy;
 import szewek.mcflux.tileentities.TileEntityEnergyMachine;
 
-public class ItemFESniffer extends Item {
+public class ItemFESniffer extends ItemMCFlux {
 	private final TextComponentTranslation
 			textBlock = new TextComponentTranslation("mcflux.blockcompat.start"),
 			textEntity = new TextComponentTranslation("mcflux.entitycompat.start"),